compute: name both dispatch rails, and reach the dispatch through Backend
`compute_exec` carried both rails' dispatch executors and their private helpers
in one 6 000-line module behind thirty-six item-level `cfg`s, with
`execute_dispatch` as a two-line switch between two functions that already had
the same neutral signature.

* `compute_exec::metal` and `compute_exec::vulkan` are the rails. The Metal one
  takes the ABI record building, the staged-texture split, the nested-dispatch
  deferred writeback, and `execute_dispatch_metal`; the Vulkan one takes the
  engine destination resolve, the SPIR-V and format adapters, the stall
  watchdog, this rail's residency policy, and `execute_dispatch_linux`.
* `Backend::execute_dispatch` replaces the switch. `compute_exec` now calls
  `backend::selected().execute_dispatch(…)` and names no rail.
* `ResidentServe` deliberately stays in the shared module. Its own doc already
  said why — "declared unconditionally although only the Vulkan backend can
  produce one, so the rails that carry the answer through a `backend-metal`
  build can still name its type" — and it is the pattern the staging loops
  depend on: one shape, one accessor, `None` on the rail with no mirror.

`execute_dispatch_nested` is not a trait method yet, and the commented reason is
the real blocker: the two rails declare two different `ComputeSession` structs
under one name, so a trait signature naming it would be two signatures. That is
the next commit.

No behavior change on either arm. Metal 1315 pass, Vulkan 2173 pass, both
serial; feature-matrix green with per-arm counts held.
